Reduce size of Details struct

+++ src/sam.hpp
@@ -46,13 +46,13 @@ enum struct CigarOps {
 /* Details about aligning a single or paired-end read */
 struct Details {
     bool nam_rescue{false}; // find_nams_rescue() was needed
-    uint64_t nams{0};  // No. of NAMs found
-    uint64_t rescue_nams{0}; // No. of NAMs found during rescue
-    uint64_t nam_inconsistent{0};
-    uint64_t mate_rescue{0}; // No. of times rescue by local alignment was attempted
-    uint64_t tried_alignment{0}; // No. of computed alignments (get_alignment or rescue_mate)
-    uint64_t gapped{0};  // No. of gapped alignments computed (in get_alignment)
-    uint64_t best_alignments{0}; // No. of best alignments with same score
+    uint32_t nams{0};  // No. of NAMs found
+    uint32_t rescue_nams{0}; // No. of NAMs found during rescue
+    uint32_t nam_inconsistent{0};
+    uint32_t mate_rescue{0}; // No. of times rescue by local alignment was attempted
+    uint32_t tried_alignment{0}; // No. of computed alignments (get_alignment or rescue_mate)
+    uint32_t gapped{0};  // No. of gapped alignments computed (in get_alignment)
+    uint32_t best_alignments{0}; // No. of best alignments with same score
 
     Details& operator+=(const Details& other) {
         nam_rescue = nam_rescue || other.nam_rescue;
